+++ mol_gnn/featurizers/graph/__init__.py
@@ -1,4 +1,4 @@
 from .base import GraphFeaturizer
-from .molgraph import Graph
+from .graph import Graph
 from .molecule import SimpleMoleculeMolGraphFeaturizer
 from .reaction import CondensedGraphOfReactionFeaturizer, CGRFeaturizer, RxnMode
